+++ console_mods/helpers.py
@@ -1,0 +1,283 @@
+import tiktoken
+
+from console_mods.config_attrs import FetchConfig
+
+import json
+import locale
+import os
+import toml
+from typing import Any, AnyStr
+from termcolor import colored
+import platform
+from simple_term_menu import TerminalMenu
+from string import ascii_lowercase
+import textwrap
+from prompt_toolkit import prompt
+from prompt_toolkit.completion import PathCompleter
+import sys
+
+import readline  # Necessary for input()
+
+
+class Helper(FetchConfig):
+    @staticmethod
+    def help_info() -> None:
+        """
+        Prints the available commands
+        """
+        commands: list[str] = [
+            "cost - Display conversation costs.",
+            "file - Submit long text from a file to the chat.",
+            "flush - Start a new conversation.",
+            "format - Format multiline pasted text before sending to the chat."
+            "save - Save the current conversation to a file.",
+            "exit - Exit the program.",
+            "",
+            "help - Display this help message.",
+            "commands - Display this list of commands.",
+        ]
+        print("You can use the following commands:")
+        print("\n".join(f"\t{command}" for command in commands))
+
+    def set_locale(self) -> None:
+        match platform.system():
+            case "Darwin":
+                locale.setlocale(locale.LC_ALL, "en_US.UTF-8")
+            case "Linux":
+                locale.setlocale(locale.LC_ALL, "en_US.utf8")
+            case _:
+                self.custom_print("warn", f"Unable to detect OS. Setting a default locale.")
+                locale.setlocale(locale.LC_ALL, '')
+
+    @staticmethod
+    def __flush_lines(lines: int) -> None:
+        for line in range(lines):
+            print("\033[F\033[K", end="")
+
+    def write_to_config(self, *args, new_value: Any) -> None:
+        with open(self.CONFIG_PATH, 'r') as file:
+            config_data: dict[str, Any] = toml.load(file)
+
+        match len(args):
+            case 2:
+                config_data[args[0]][args[1]] = new_value
+            case 3:
+                config_data[args[0]][args[1]][args[2]] = new_value
+            case _:
+                self.custom_print("error", "Wrong usage of write_to_config", 1)
+
+        with open('config.toml', 'w') as file:
+            toml.dump(config_data, file)
+
+    def __role_preview(self, item: str) -> str:
+        rows, columns = os.popen('stty size', 'r').read().split()
+        line_length: int = int(columns) // 2
+        match item:
+            case "Add New system behavior":
+                return "Write down how you would like the AI to act like."
+            case "Exit":
+                return "Terminate the application."
+            case "Default":
+                return '\n'.join(textwrap.wrap(self.ALL_ROLES.get(self.DEFAULT_ROLE), width=line_length))
+            case _:
+                return '\n'.join(textwrap.wrap(self.ALL_ROLES.get(item, "Unknown Option"), width=line_length))
+
+    def base_chat_menu(self, title: str, default_option: str | list, base_options: list,
+                       add_nums: bool = True, preview_func: Any = None) -> str:
+        """
+        Base terminal menu
+        :param title: Title of the terminal menu
+        :param default_option: Place the items which you would like to appear first here
+        :param base_options: The available "clickable" options
+        :param add_nums: By default if the options are < 10 they will be numerated
+        :param preview_func: asd
+        If set to False, they will have alphabetic ordering.
+        :return: The selected string from the menu.
+        """
+        enum_options: list[str] = []
+        counter: int = 1
+        letters_counter: int = 0
+        default_option: str | list = [default_option] if isinstance(default_option, str) else default_option
+        options: list[str] = default_option + base_options + ["Exit"]
+        for opt in options:
+            match opt:
+                case "Skip":
+                    enum_options.append("[s] Skip")
+                case "Exit":
+                    enum_options.append("[x] Exit")
+                case _:
+                    if add_nums and len(options) < 10:
+                        enum_options.append("[{}] {}".format(counter, opt))
+                        counter += 1
+                    else:
+                        letters: list[str] = [x for x in list(ascii_lowercase) if x not in ["s", "x"]]
+                        enum_options.append("[{}]{}".format(letters[letters_counter], opt))
+                        letters_counter += 1
+        if preview_func:
+            terminal_menu: Any = TerminalMenu(enum_options, title=title, preview_command=preview_func)
+        else:
+            terminal_menu: Any = TerminalMenu(enum_options, title=title)
+        menu_entry_index: Any = terminal_menu.show()
+        if menu_entry_index is None:
+            self.custom_print("error", "Keyboard interrupt or attempt to break stuff ;(", 130)
+        selected_item: str = options[menu_entry_index]
+        if selected_item == "Exit":
+            self.custom_print("info", "Goodbye :)", 0, True)
+        return selected_item
+
+    def select_temperature(self) -> None:
+        """
+        Handles the chat temperature (randomness).
+        """
+        if self.SHOW_TEMPERATURE_PICKER:
+            user_input: str = ""
+            lines: int = 1
+            while True:
+                try:
+                    user_input = self.custom_input(f"Press 'ENTER' for the default setting ({self.CHAT_TEMPERATURE}): ")
+                    float_temp: float = float(user_input)
+                    if 2 >= float_temp >= 0:
+                        self.__flush_lines(lines)
+                        self.CHAT_TEMPERATURE = float_temp
+                        break
+                except ValueError:
+                    if not user_input:
+                        self.__flush_lines(lines)
+                        break
+                lines += 1
+
+    def __add_custom_role(self) -> str | None:
+        try:
+            while True:
+                role_title: str = self.custom_input("Enter a title for the new role: ")
+                if not role_title:
+                    self.custom_print("warn", "Please fill the title!")
+                    continue
+                if role_title not in self.ALL_ROLES.keys():
+                    break
+                self.custom_print("warn", "Such role name already exists!")
+                continue
+
+            while True:
+                role_desc: str = self.custom_input("Enter a detailed description of your custom role: ")
+                if role_desc:
+                    break
+                self.custom_print("warn", "Please fill the description!")
+                continue
+
+            self.write_to_config("chat", "roles", role_title, new_value=role_desc)
+            return role_desc
+        except KeyboardInterrupt:
+            return self.ALL_ROLES.get(self.DEFAULT_ROLE)
+
+    def roles_chat_menu(self) -> str:
+        """
+        Handle the roles within the config file.
+        If you don't like them, you can create your own.
+        """
+        if self.SHOW_ROLE_SELECTION:
+            roles_names: list[str] = list(self.ALL_ROLES.keys())
+            roles_names.remove(self.DEFAULT_ROLE)
+            roles_names.append("Add New system behavior")
+            selected_role = self.base_chat_menu(
+                f'Select a role or use the default one "{self.DEFAULT_ROLE}":',
+                "Default",
+                roles_names,
+                add_nums=False,
+                preview_func=self.__role_preview
+            )
+            if selected_role == "Add New system behavior":
+                role: str = self.__add_custom_role()
+            else:
+                role: str = self.ALL_ROLES.get(selected_role)
+            if not role:
+                return self.ALL_ROLES.get(self.DEFAULT_ROLE)
+            return role
+        return self.ALL_ROLES.get(self.DEFAULT_ROLE)
+
+    def continue_chat_menu(self) -> list[dict[str, str]] | None:
+        """
+        Given a Base Path craws a subdir called "chats"
+        If there are no files there, it will automatically skip
+        the whole function.
+        :return: a string which is later handled by continue_chat()
+        """
+        all_chats: list[str] = os.listdir(self.CHATS_PATH)
+        if not len(all_chats):
+            return None
+        selection: str = self.base_chat_menu("Would you like to continue a previous chat?:", "Skip", all_chats)
+        if selection == "Skip":
+            return None
+        try:
+            full_path: str = os.path.join(self.CHATS_PATH, selection)
+            with open(full_path, "r") as file:
+                data: Any = json.load(file)
+                self.custom_print("ok", f"Successfully loaded previous chat - {selection}")
+            return data
+        except json.JSONDecodeError as e:
+            self.custom_print("error", f"Error decoding JSON: {e}", 1)
+
+    def file_prompt(self) -> str | None:
+        """
+        Opens a custom prompt which supports tab completion.
+        If a file is selected that path is being returned.
+        If the function is aborted a False is returned.
+        """
+        try:
+            while True:
+                user_input = prompt("Enter a path: ", completer=PathCompleter())
+                if not os.path.isfile(user_input):
+                    self.custom_print("warn", f"No such file at - {user_input}")
+                    continue
+                with open(user_input, "r") as file:
+                    user_prompt: AnyStr = file.read()
+                    user_prompt.replace("\n", "\\n").replace('"', '\\"')
+                    context: str = input(
+                        colored(
+                            "Add any additional clarification before the submitted text or press 'ENTER' to continue: ",
+                            "blue",
+                        )
+                    )
+                    if context:
+                        user_prompt = context + ":\n" + user_prompt
+                return user_prompt
+        except KeyboardInterrupt:
+            self.custom_print("info", "Cancelled the file selection, continuing with the chat.")
+            return None
+
+    def format_multiline(self):
+        """
+        Formats a multiline chat input.
+        """
+        try:
+            self.custom_print("info", "Paste the multiline text and press 'Ctrl+D' on an new empty line to continue: ")
+            content = sys.stdin.read()
+            if content:
+                content.replace("\n", "\\n").replace('"', '\\"')
+                context = input(
+                    colored(
+                        "Add additional clarification before the formatted text or press 'ENTER' to continue: ",
+                        "blue"))
+                if context:
+                    content = context + ":\n" + content
+            return content
+        except KeyboardInterrupt:
+            print("\b\b", end="")
+            self.custom_print("info", "Cancelled the multiline text, continuing with the chat.")
+
+    def num_tokens_from_messages(self, messages: list[dict]) -> int:
+        """
+        Count the tokens of the next user prompt
+        """
+        encoding = tiktoken.encoding_for_model(self.CHAT_MODEL)
+        tokens_per_message = 3
+        tokens_per_name = 1
+        num_tokens = 0
+        for message in messages:
+            num_tokens += tokens_per_message
+            for key, value in message.items():
+                num_tokens += len(encoding.encode(value))
+                if key == "name":
+                    num_tokens += tokens_per_name
+        num_tokens += 3
+        return num_tokens
