+++ requirements.txt
@@ -1,7 +1,7 @@
-openai~=0.27.8
-termcolor~=1.1.0
-toml~=0.10.2
-prompt-toolkit~=3.0.39
-simple_term_menu~=1.6.1
-Pygments~=2.15.1
-tiktoken~=0.4.0
+openai==0.27.8
+prompt_toolkit==3.0.39
+Pygments==2.16.1
+simple_term_menu==1.6.1
+termcolor==2.3.0
+tiktoken==0.4.0
+toml==0.10.2
